+++ gaffer-as-a-service/gaas-rest/src/main/java/uk/gov/gchq/gaffer/gaas/client/GafferClient.java
@@ -28,8 +28,8 @@ import org.springframework.stereotype.Repository;
 import uk.gov.gchq.gaffer.gaas.exception.GaaSRestApiException;
 import uk.gov.gchq.gaffer.gaas.handlers.DeploymentHandler;
 import uk.gov.gchq.gaffer.gaas.model.GaaSGraph;
-import uk.gov.gchq.gaffer.gaas.model.Gaffer;
 import uk.gov.gchq.gaffer.gaas.model.GraphUrl;
+import uk.gov.gchq.gaffer.gaas.model.v1.Gaffer;
 
 import java.util.List;
 import java.util.stream.Collectors;
