+++ gaffer-as-a-service/gaas-rest/src/main/java/uk/gov/gchq/gaffer/gaas/model/v1/Gaffer.java
@@ -14,39 +14,43 @@
  * limitations under the License.
  */
 
-package uk.gov.gchq.gaffer.gaas.model;
+package uk.gov.gchq.gaffer.gaas.model.v1;
 
 import com.google.gson.annotations.SerializedName;
-import io.kubernetes.client.common.KubernetesListObject;
 import io.kubernetes.client.common.KubernetesObject;
-import io.kubernetes.client.openapi.models.V1ListMeta;
+import io.kubernetes.client.openapi.models.V1ObjectMeta;
 import org.apache.commons.lang3.builder.EqualsBuilder;
 import org.apache.commons.lang3.builder.HashCodeBuilder;
 import org.apache.commons.lang3.builder.ToStringBuilder;
 import uk.gov.gchq.gaffer.gaas.util.Constants;
 
-import java.util.List;
 
-public class GafferList implements KubernetesListObject {
-
-    @SerializedName(Constants.SERIALISED_NAME_METADATA)
-    private V1ListMeta metadata;
-
-    @SerializedName(Constants.SERIALISED_NAME_ITEMS)
-    private List<Gaffer> items;
+/**
+ * Value object for Gaffer
+ */
+public class Gaffer implements KubernetesObject {
 
     @SerializedName(Constants.SERIALISED_NAME_API_VERSION)
     private String apiVersion;
 
+
     @SerializedName(Constants.SERIALISED_NAME_KIND)
     private String kind;
 
-    public V1ListMeta getMetadata() {
-        return metadata;
-    }
+    @SerializedName(Constants.SERIALISED_NAME_METADATA)
+    private V1ObjectMeta metadata;
 
-    public List<? extends KubernetesObject> getItems() {
-        return items;
+
+    @SerializedName(Constants.SERIALISED_NAME_SPEC)
+    private GafferSpec spec;
+
+
+    @SerializedName(Constants.SERIALISED_NAME_STATUS)
+    private GafferStatus status;
+
+
+    public V1ObjectMeta getMetadata() {
+        return metadata;
     }
 
     public String getApiVersion() {
@@ -55,6 +59,39 @@ public class GafferList implements KubernetesListObject {
 
     public String getKind() {
         return kind;
+    }
+
+    public GafferSpec getSpec() {
+        return spec;
+    }
+
+    public GafferStatus getStatus() {
+        return status;
+    }
+
+    public Gaffer status(final GafferStatus status) {
+        this.status = status;
+        return this;
+    }
+
+    public Gaffer metaData(final V1ObjectMeta metadata) {
+        this.metadata = metadata;
+        return this;
+    }
+
+    public Gaffer apiVersion(final String apiVersion) {
+        this.apiVersion = apiVersion;
+        return this;
+    }
+
+    public Gaffer kind(final String kind) {
+        this.kind = kind;
+        return this;
+    }
+
+    public Gaffer spec(final GafferSpec spec) {
+        this.spec = spec;
+        return this;
     }
 
     @Override
@@ -67,33 +104,36 @@ public class GafferList implements KubernetesListObject {
             return false;
         }
 
-        final GafferList that = (GafferList) o;
+        final Gaffer gaffer = (Gaffer) o;
 
         return new EqualsBuilder()
-                .append(metadata, that.metadata)
-                .append(items, that.items)
-                .append(apiVersion, that.apiVersion)
-                .append(kind, that.kind)
+                .append(apiVersion, gaffer.apiVersion)
+                .append(kind, gaffer.kind)
+                .append(metadata, gaffer.metadata)
+                .append(spec, gaffer.spec)
+                .append(status, gaffer.status)
                 .isEquals();
     }
 
     @Override
     public int hashCode() {
         return new HashCodeBuilder(17, 37)
-                .append(metadata)
-                .append(items)
                 .append(apiVersion)
                 .append(kind)
+                .append(metadata)
+                .append(spec)
+                .append(status)
                 .toHashCode();
     }
 
     @Override
     public String toString() {
         return new ToStringBuilder(this)
-                .append("metadata", metadata)
-                .append("items", items)
                 .append("apiVersion", apiVersion)
                 .append("kind", kind)
+                .append("metadata", metadata)
+                .append("spec", spec)
+                .append("status", status)
                 .toString();
     }
 }
