+++ gaffer-as-a-service/gaas-rest/src/test/java/uk/gov/gchq/gaffer/gaas/model/v1/GafferSpecTest.java
@@ -15,7 +15,7 @@
  */
 
 
-package uk.gov.gchq.gaffer.gaas.model;
+package uk.gov.gchq.gaffer.gaas.model.v1;
 
 import org.junit.jupiter.api.Test;
