+++ gaffer-as-a-service/gaas-rest/src/test/java/uk/gov/gchq/gaffer/gaas/services/GetGaffersServiceTest.java
@@ -26,7 +26,7 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import uk.gov.gchq.gaffer.gaas.client.GafferClient;
 import uk.gov.gchq.gaffer.gaas.exception.GaaSRestApiException;
 import uk.gov.gchq.gaffer.gaas.model.GaaSGraph;
-import uk.gov.gchq.gaffer.gaas.model.RestApiStatus;
+import uk.gov.gchq.gaffer.gaas.model.v1.RestApiStatus;
 import uk.gov.gchq.gaffer.gaas.util.UnitTest;
 
 import java.util.ArrayList;
